+++ packages/typespec-go/test/local/armtest/zz_bodyroots_client.go
@@ -89,6 +89,7 @@ func (client *BodyRootsClient) actionCreateRequest(ctx context.Context, resource
 	}
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2025-01-01")
+	reqQP.Set("filter", "constantValue")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, action); err != nil {